Implementation PCA9685 board

+++ wiredbot_teleop/scripts/wiredbot_teleop_x8r_node.py
@@ -1,7 +1,7 @@
 #! /usr/bin/env python
 
 import rospy
-from std_msgs.msg import Int32
+from std_msgs.msg import Int8
 from geometry_msgs.msg import Twist
 
 goal_data = 0
@@ -11,7 +11,7 @@ _cmd_vel_pub = rospy.Publisher('/cmd_vel', Twist, queue_size=2)
 def move_cmd(direction="-"):
 
     rospy.init_node('wiredbot_teleop_x8r')
-    _cmd_vel_sub = rospy.Subscriber('/teleop_ch3', Int32, callback_wiredbot_teleop)
+    _cmd_vel_sub = rospy.Subscriber('/teleop_ch3', Int8, callback_wiredbot_teleop)
     _twist_object = Twist()
 
     rate = rospy.Rate(2)
